split tests into fs and browser
